Feat: 이미지 업로드 기능 추가 및 태그 관련 기능 삭제

+++ src/main/java/com/example/hyfit_server/service/post/PostService.java
@@ -1,14 +1,9 @@
 package com.example.hyfit_server.service.post;
 
 import com.example.hyfit_server.config.response.BaseException;
-import com.example.hyfit_server.config.response.BaseResponseStatus;
-import com.example.hyfit_server.domain.post.ImageRepository;
-import com.example.hyfit_server.domain.post.PostEntity;
-import com.example.hyfit_server.domain.post.PostRepository;
-import com.example.hyfit_server.domain.post.PostTagMapRepository;
-import com.example.hyfit_server.dto.Post.PostDto;
-import com.example.hyfit_server.dto.Post.PostModifyDto;
-import com.example.hyfit_server.dto.Post.PostSaveDto;
+import com.example.hyfit_server.domain.post.*;
+import com.example.hyfit_server.dto.Image.ImageSaveDto;
+import com.example.hyfit_server.dto.Post.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -26,21 +21,31 @@ import static com.example.hyfit_server.config.response.BaseResponseStatus.*;
 public class PostService {
 
     private final PostRepository postRepository;
-    private final PostTagMapRepository postTagMapRepository;
     private final ImageRepository imageRepository;
 
-    public PostDto savePost(PostSaveDto postSaveDto, Errors errors) throws BaseException {
+    public PostSaveResDto savePost(PostSaveDto postSaveDto, String imageUrl, Errors errors) throws BaseException {
         if(errors.hasErrors()) {
-            throw new BaseException(NO_POST_TITLE);
+            throw new BaseException(NO_POST_CONTENTS);
         }
-        PostEntity postResult = postRepository.save(postSaveDto.toEntity());
-        return postResult.toDto();
+        PostSaveResDto result = new PostSaveResDto();
+        PostEntity postEntity = postRepository.save(postSaveDto.toEntity());
+        result.setPostDto(postEntity.toDto());
+
+        if (imageUrl != null) {
+            ImageSaveDto imageSaveDto = ImageSaveDto.builder()
+                    .postId(postEntity.getPostId())
+                    .imageUrl(imageUrl)
+                    .build();
+
+            ImageEntity imageEntity = imageRepository.save(imageSaveDto.toEntity());
+            result.setImageDto(imageEntity.toDto());
+        } else {
+            result.setImageDto(null);
+        }
+        return result;
     }
 
-    public List<PostDto> getAllPosts(String email) throws BaseException {
-        if(postRepository.findAllByEmail(email).size() == 0) {
-            throw new BaseException(NO_SAVED_POST);
-        }
+    public List<PostDto> getAllPostsOfUser(String email) throws BaseException {
         List<PostDto> result = postRepository.findAllByEmail(email)
                 .stream().map(m -> m.toDto())
                 .collect(Collectors.toList());
@@ -49,31 +54,33 @@ public class PostService {
 
     public PostDto getOnePost(String email, long id) throws BaseException {
         PostEntity postEntity = postRepository.findByEmailAndPostId(email, id);
+        if(postEntity == null) {
+            throw new BaseException(NO_POST_ERROR);
+        }
         return postEntity.toDto();
     }
 
-    public PostDto modify(String email, long id, PostModifyDto postModifyDto) throws BaseException {
+    public PostDto modifyPost(String email, long id, PostModifyDto postModifyDto) throws BaseException {
         PostEntity postEntity = postRepository.findByEmailAndPostId(email, id);
-//        if(postModifyDto.getContent() == null) {
-//            postModifyDto.setContent(postEntity.getContent());
-//        }
-//        postEntity.modify(postModifyDto);
-        postModifyDto.getTitle().ifPresent(postEntity::modifyTitle);
+        if(postEntity == null) {
+            throw new BaseException(NO_POST_ERROR);
+        }
+
         postModifyDto.getContent().ifPresent(postEntity::modifyContent);
 
         return postEntity.toDto();
     }
 
-    public void deletePost(long id) throws BaseException {
-        PostEntity postEntity = postRepository.findByPostId(id);
+    public void deletePost(String email, long id) throws BaseException {
+        PostEntity postEntity = postRepository.findByEmailAndPostId(email, id);
+        if(postEntity == null) {
+            throw new BaseException(NO_POST_ERROR);
+        }
 
         // image 삭제
         imageRepository.deleteAllByPostId(id);
 
-        // post_tag_map 삭제
-        postTagMapRepository.deleteAllByPostId(id);
-
         postRepository.delete(postEntity);
-
     }
+
 }
